+++ x/leverage/client/tests/suite.go
@@ -75,7 +75,7 @@ func runTestTransactions(s *IntegrationTestSuite, tcs []testTransaction) {
 			s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), resp), out.String())
 
 			if tc.expectedErr == nil {
-				s.Require().Equal(0, int(resp.Code))
+				s.Require().Equal(0, int(resp.Code), "events %v", resp.Events)
 			} else {
 				s.Require().Equal(int(tc.expectedErr.ABCICode()), int(resp.Code))
 			}
@@ -143,6 +143,9 @@ func (s *IntegrationTestSuite) TestQueryAllRegisteredTokens() {
 						MaxBorrowRate:        sdk.MustNewDecFromStr("1.5"),
 						KinkUtilizationRate:  sdk.MustNewDecFromStr("0.2"),
 						LiquidationIncentive: sdk.MustNewDecFromStr("0.18"),
+						EnableMsgLend:        true,
+						EnableMsgBorrow:      true,
+						Blacklist:            false,
 					},
 				},
 			},
@@ -1063,13 +1066,13 @@ func (s *IntegrationTestSuite) TestQueryBorrowLimit() {
 	runTestTransactions(s, cleanupCommands)
 }
 
-func (s *IntegrationTestSuite) TestQueryLiquidationLimit() {
+func (s *IntegrationTestSuite) TestQueryLiquidationThreshold() {
 	val := s.network.Validators[0]
 
 	simpleCases := []testQuery{
 		{
 			"invalid address",
-			cli.GetCmdQueryLiquidationLimit(),
+			cli.GetCmdQueryLiquidationThreshold(),
 			[]string{
 				"xyz",
 			},
@@ -1078,15 +1081,15 @@ func (s *IntegrationTestSuite) TestQueryLiquidationLimit() {
 			nil,
 		},
 		{
-			"query zero liquidation limit",
-			cli.GetCmdQueryLiquidationLimit(),
+			"query zero liquidation threshold",
+			cli.GetCmdQueryLiquidationThreshold(),
 			[]string{
 				val.Address.String(),
 			},
 			false,
-			&types.QueryLiquidationLimitResponse{},
-			&types.QueryLiquidationLimitResponse{
-				LiquidationLimit: sdk.ZeroDec(),
+			&types.QueryLiquidationThresholdResponse{},
+			&types.QueryLiquidationThresholdResponse{
+				LiquidationThreshold: sdk.ZeroDec(),
 			},
 		},
 	}
@@ -1115,18 +1118,18 @@ func (s *IntegrationTestSuite) TestQueryLiquidationLimit() {
 
 	nonzeroCase := []testQuery{
 		{
-			"query nonzero liquidation limit",
-			cli.GetCmdQueryLiquidationLimit(),
+			"query nonzero liquidation threshold",
+			cli.GetCmdQueryLiquidationThreshold(),
 			[]string{
 				val.Address.String(),
 			},
 			false,
-			&types.QueryLiquidationLimitResponse{},
-			&types.QueryLiquidationLimitResponse{
+			&types.QueryLiquidationThresholdResponse{},
+			&types.QueryLiquidationThresholdResponse{
 				// From app/test_helpers.go/IntegrationTestNetworkConfig
 				// This result is umee's liquidation threshold times the collateral
 				// amount loaned, times its initial oracle exchange rate.
-				LiquidationLimit: sdk.MustNewDecFromStr("34.21"),
+				LiquidationThreshold: sdk.MustNewDecFromStr("34.21"),
 				// 0.05 * 20 * 34.21 = 34.21
 			},
 		},
@@ -1551,7 +1554,7 @@ func (s *IntegrationTestSuite) TestCmdBorrow() {
 				val.Address.String(),
 				"70uumee",
 			},
-			types.ErrBorrowLimitLow,
+			types.ErrUndercollaterized,
 		},
 		{
 			"borrow",
@@ -1716,22 +1719,22 @@ func (s *IntegrationTestSuite) TestCmdRepay() {
 
 	testCases := []testTransaction{
 		{
-			"invalid asset",
+			"nothing to repay",
 			cli.GetCmdRepayAsset(),
 			[]string{
 				val.Address.String(),
 				"10xyz",
 			},
-			types.ErrInvalidAsset,
+			types.ErrInvalidRepayment,
 		},
 		{
-			"invalid asset (uToken)",
+			"nothing to repay (uToken)",
 			cli.GetCmdRepayAsset(),
 			[]string{
 				val.Address.String(),
 				"10u/umee",
 			},
-			types.ErrInvalidAsset,
+			types.ErrInvalidRepayment,
 		},
 		{
 			"repay",